+++ src/components/Card/Card.css
@@ -1,21 +1,3 @@
-.wrap {
-    padding: 22px;
-}
-
-.cards {
-    margin: 0 auto;
-    display: flex;
-    flex-wrap: wrap;
-    text-align: center;
-    justify-content: space-between;
-    width: 900px;
-}
-
-.image_card {
-    width: 200px;
-    padding: 0 0 20px;
-}
-
 .container_card {
     width: 250px;
     height: 350px;
