+++ src/components/Input/input.js
@@ -4,13 +4,11 @@ import { ImEye } from "react-icons/im";
 import { ImEyeBlocked } from "react-icons/im";
 
 const InputLogin = (props) => {
-    const [isInputButton, setInputButton] = useState(false);
     const [isPasswordShown, setPasswordShow] = useState(true);
 
     const buttonClick = () => {
-        setInputButton(prevState => !prevState);
         setPasswordShow(prevState => !prevState);
-    }
+    };
 
     return (
         <div>
@@ -18,7 +16,7 @@ const InputLogin = (props) => {
                 placeholder={props.isLogin ? "User Name" : "Password"}
                 type={props.isPassword && isPasswordShown ? 'password' : 'text'} />
             <button onClick={buttonClick} className='input_button'>
-                {isInputButton ? <ImEye className='eaeOpen' /> : <ImEyeBlocked className='eaeClose' />}
+                {isPasswordShown ? <ImEyeBlocked className='eaeClose' /> : <ImEye className='eaeOpen' />}
             </button >
         </div>
     );
